🔧 refactor: abstract extract colors function into utils
this function has nothing to do with any internal state
